# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learn Hub",
        "description": "API doc for Learn Hub",
        "contact": {
            "url": "https://justice-chimobi.vercel.app/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://learn-hub-backend.test/api/v1"
        }
    ],
    "paths": {
...
        "/threads/edit/{thread}": {
            "patch": {
                "tags": [
                    "threads"
                ],
                "summary": "Update thread created by a author",
                "description": "Update thread created by a author",
                "operationId": "9195ddf2cc4905449293bf496efae5ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Thread ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ID": {
                                "summary": "An UUID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update thread created by a author",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "learn-hub thread name"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "content of the thread body"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Thread updated successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "You are not authorized to edit this thread."
                    },
                    "404": {
                        "description": "Thread not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/threads/{thread}": {
            "get": {
                "tags": [
                    "threads"
                ],
                "summary": "Get an thread by ID",
                "description": "Get an thread by ID",
                "operationId": "1ce2ac18dc33cc3e75f591c11a6dd181",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Thread ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ThreadResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Thread not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/threads/{thread}/likes": {
            "post": {
                "tags": [
                    "threads"
                ],
                "summary": "like a thread",
                "description": "user like can only like a thread if not liked before.",
                "operationId": "b768269e5542a2d5e86064513b2f5004",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Thread ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ID": {
                                "summary": "An UUID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/threads/{thread}/dislikes": {
            "delete": {
                "tags": [
                    "threads"
                ],
                "summary": "dislike a thread",
                "description": "user dislike his/her liked thread",
                "operationId": "6e44a0b83dfc74b44c698f7ac7cec1e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Thread ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ID": {
                                "summary": "An UUID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/accounts/upload-avatar": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update current authenticated user avatar profile, please ensure you get your image url from upload tag image/upload endpoint",
                "description": "Update current authenticated user avatar profile",
                "operationId": "1cc6412bb16981714fdfd807f6d36659",
                "requestBody": {
                    "description": "Update user avatar profile",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "example": "https://res.cloudinary.com/dbx3dhfkt/image/upload/v1672045944/estudy/pictures/image-5a9482cd3-a97e-4627-dbc3-9cb53797e40a.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Updated successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "401": {
                        "description": "Unathenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/accounts/delete": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete user account",
                "description": "Delete user account",
                "operationId": "8f5af3b81be73911ae13d25d717f5822",
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DateTimeResource": {
                "title": "DateTimeResource",
                "description": "format date time",
                "properties": {
                    "human": {
                        "description": "long format string",
                        "type": "string",
                        "example": "2 hours ago"
                    },
                    "date_time": {
                        "description": "string representation",
                        "type": "string",
                        "example": "2024-01-22 12:01:58"
                    },
                    "human_short": {
                        "description": "short format string",
                        "type": "string",
                        "example": "2h ago"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DateTimeResource"
                }
            },
...
            "ThreadResource": {
                "title": "ThreadResource",
                "description": "Thread Resource",
                "properties": {
                    "id": {
                        "description": "Course ID",
                        "type": "string",
                        "example": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                    },
                    "title": {
                        "description": "Article title",
                        "type": "string",
                        "example": "Article intro"
                    },
                    "slug": {
                        "description": "Article slug title",
                        "type": "string",
                        "example": "article-intro"
                    },
                    "content": {
                        "description": "Article content",
                        "type": "string",
                        "example": "About my article content."
                    },
                    "thread_comment_counts": {
                        "description": "Number of Thread comments",
                        "type": "number",
                        "example": "67"
                    },
                    "thread_like_counts": {
                        "description": "Number of Thread likes",
                        "type": "number",
                        "example": "47"
                    },
                    "user_liked_thread": {
                        "description": "if currently authenticated user like a thread already",
                        "type": "boolean",
                        "example": "false"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/DateTimeResource"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserShortResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ThreadResource"
                }
            },
...
            "UserShortResource": {
                "title": "UserShortResource",
                "description": "User article Resource",
                "properties": {
                    "id": {
                        "description": "Id",
                        "type": "string",
                        "example": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                    },
                    "fullname": {
                        "description": "Full name",
                        "type": "string",
                        "example": "Justice Chimobi"
                    },
                    "username": {
                        "description": "Username",
                        "type": "string",
                        "example": "justice-chimobi"
                    },
                    "avatar": {
                        "description": "Course name",
                        "type": "string",
                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg"
                    },
                    "profile_headlines": {
                        "description": "profile_headlines",
                        "type": "string",
                        "example": "Frontend Developer || Full Stack"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserShortResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}